fix: 修复数据字段不一致

+++ app/frontend/biz/service/add_cart_item.go
@@ -39,7 +39,7 @@ func (h *AddCartItemService) Run(req *cart.AddCartReq) (resp *common.Empty, err 
 		UserId: frontendutils.GetUserIdFromCtx(h.Context),
 		Item: &rpccart.CartItem{
 			ProductId: req.ProductId,
-			Quantity:  req.ProductNum,
+			Quantity:  uint32(req.ProductNum),
 		},
 	})
 	return
